+++ src/command_modules/azure-cli-interactive/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+0.3.18
+++++++
+* Completions kick in as soon as command table loading is done.
+* Fix bug with using `--style` parameter.
+
 0.3.17
 ++++++
 * Persist history across different sessions