+++ src/command_modules/azure-cli-lab/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.0.18
+++++++
+* Minor fixes.
+
 0.0.17
 ++++++
 * Minor fixes.