+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -2,8 +2,14 @@
 
 Release History
 ===============
+
+2.0.29
+++++++
+* Minor fixes. 
+
 2.0.28
 ++++++
+* vm/vmss create: support to attach unmanaged data disks and configure their caching modes 
 * vm/vmss: author managed identity commands `identity assign/remove/show`, and deprecate `assign-identity/remove-identity`
 * vmss create: default priority to None
 * Support Autorest 3.0 based SDKs